requirements.txt
asgiref==3.8.1
asttokens==2.4.1
black==24.4.2
celery==5.4.0
click==8.1.7
colorama==0.4.6
coverage==7.5.1
CurrencyConverter==0.17.21
decorator==5.1.1
Django==4.2.9
django-dump-load-utf8==0.0.4
django-filter==24.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
executing==2.0.1
flake8==7.0.0
forex-python==1.8
inflection==0.5.1
ipython==8.24.0
isort==5.13.2
jedi==0.19.1
matplotlib-inline==0.1.7
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.0
parso==0.8.4
pathspec==0.12.1
pillow==10.3.0
platformdirs==4.2.1
prompt-toolkit==3.0.43
psycopg2==2.9.9
pure-eval==0.2.2
pycodestyle==2.11.1
pyflakes==3.2.0
Pygments==2.18.0
PyJWT==2.8.0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
redis==5.0.4
setuptools==69.5.1
six==1.16.0
sqlparse==0.5.0
stack-data==0.6.3
stripe==9.6.0
traitlets==5.14.3
tzdata==2024.1
uritemplate==4.1.1
wcwidth==0.2.13
